added route maker and modified request class

+++ app/src/main/java/com/thesis/ashline/localnewsscraper/model/Search.java
@@ -4,11 +4,11 @@ package com.thesis.ashline.localnewsscraper.model;
  * Created by ashfire on 1/16/15.
  */
 public class Search {
-    public int id;
+    public long id;
     public String latitude;
     public String longitude;
     public String date;
-    public int category_id;
+    public long category_id;
     //todo public int place_id
 
 }
